+++ src/RestApi/Device/Authentication/DeviceAuthenticationManagement.ts
@@ -10,11 +10,21 @@ export default class DeviceAuthenticationManagement {
 		return `${DEVICE}/${deviceUid}/authentication`;
 	}
 
+	private static getUrlByAuthHash(authHash: string): string {
+		return `${DEVICE}/authentication/${authHash}`;
+	}
+
 	constructor(private options: IOptions) {
 	}
 
 	public async get(deviceUid: string): Promise<IDeviceAuthentication> {
 		const response = await getResource(this.options, DeviceAuthenticationManagement.getUrl(deviceUid));
+
+		return new DeviceAuthentication(await parseJSONResponse(response));
+	}
+
+	public async getByAuthHash(authHash: string): Promise<IDeviceAuthentication> {
+		const response = await getResource(this.options, DeviceAuthenticationManagement.getUrlByAuthHash(authHash));
 
 		return new DeviceAuthentication(await parseJSONResponse(response));
 	}